fix(tests): resolve build errors in C# test project
- Fix method name typos with spaces (CS1002 errors)

- Add unique prefixes to duplicate helper class definitions

- Fix type conversion and FluentAssertions syntax errors

- Fix array property access and Task.WhenAll usage

+++ QA Tests/C# Tests/UNOPS.PAO.Business.Tests/OpportunitySections/LoadTests.cs
@@ -320,7 +320,7 @@ namespace UNOPS.PAO.Business.Tests.OpportunitySections
             await Task.WhenAll(tasks);
 
             // Assert
-            failedCount.Should().BeLessThan(requestsOverPoolSize * 0.1, "Less than 10% should fail");
+            failedCount.Should().BeLessThan((int)(requestsOverPoolSize * 0.1), "Less than 10% should fail");
             // Queueing is acceptable, complete failures are not
         }
